backend/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0
pydantic-settings
sqlalchemy[asyncio]
asyncpg
pgvector
sentence-transformers
groq
httpx
loguru
python-dotenv
python-multipart
tenacity
beautifulsoup4
lxml
apscheduler

